test: house Common Lisp tests with the other language extractors
Move the Common Lisp extraction tests into the language-extractor suite
where the project keeps per-language coverage, so they sit with their peers
rather than in the general multi-language file.

🄯 Brian O'Reilly <fade@deepsky.com>, 2026
