+++ tests/test_multilang.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import shutil
 from pathlib import Path
 import pytest
-from graphify.extract import extract_js, extract_go, extract_rust, extract, extract_sql, extract_commonlisp
+from graphify.extract import extract_js, extract_go, extract_rust, extract, extract_sql
 
 FIXTURES = Path(__file__).parent / "fixtures"
 
@@ -487,258 +487,3 @@ def test_sql_schema_qualified_alter_fk():
     for e in fk_edges:
         assert e["source"] in node_ids, f"dangling source: {e['source']}"
         assert e["target"] in node_ids, f"dangling target: {e['target']}"
-
-
-# ── Common Lisp ──────────────────────────────────────────────────────────────
-
-def test_cl_finds_package():
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    assert "error" not in r
-    assert "http-server" in _labels(r)
-
-def test_cl_finds_class():
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    assert "server" in _labels(r)
-    assert "ssl-server" in _labels(r)
-
-def test_cl_finds_defun():
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    labels = _labels(r)
-    assert any("make-server" in l for l in labels)
-    assert any("start" in l for l in labels)
-    assert any("stop" in l for l in labels)
-
-def test_cl_finds_generic():
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    assert any("process-request" in l for l in _labels(r))
-
-def test_cl_finds_macro():
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    assert any("with-server" in l and "macro" in l for l in _labels(r))
-
-def test_cl_emits_calls():
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    calls = _call_pairs(r)
-    # start() calls process-request
-    assert any("start" in src and "process-request" in tgt for src, tgt in calls)
-    # with-server macro calls make-server and start
-    assert any("with-server" in src and "make-server" in tgt for src, tgt in calls)
-    assert any("with-server" in src and "start" in tgt for src, tgt in calls)
-
-def test_cl_calls_are_extracted():
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    for e in r["edges"]:
-        if e["relation"] == "calls":
-            assert e["confidence"] == "EXTRACTED"
-
-def test_cl_no_dangling_edges():
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    node_ids = {n["id"] for n in r["nodes"]}
-    for e in r["edges"]:
-        if e["relation"] in ("contains", "method", "calls"):
-            assert e["source"] in node_ids
-
-def test_cl_docstrings():
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    rationale_edges = [e for e in r["edges"] if e["relation"] == "rationale_for"]
-    assert len(rationale_edges) >= 3  # make-server, start, process-request have docstrings
-    labels = _labels(r)
-    assert any("Process an incoming" in l for l in labels)
-
-def test_cl_method_specializers():
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    spec_edges = [e for e in r["edges"] if e["relation"] == "specializes"]
-    assert len(spec_edges) >= 1
-    # process-request specializes on server
-    assert any("process_request" in e["source"] and "server" in e["target"] for e in spec_edges)
-
-def test_cl_inherits():
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    inherit_edges = [e for e in r["edges"] if e["relation"] == "inherits"]
-    assert len(inherit_edges) >= 1
-    assert any("ssl_server" in e["source"] and "server" in e["target"] for e in inherit_edges)
-
-def test_cl_imports():
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    import_edges = [e for e in r["edges"] if e["relation"] == "imports"]
-    targets = {e["target"] for e in import_edges}
-    assert "cl" in targets
-    assert "alexandria" in targets
-
-def test_cl_finds_deftype():
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    assert "port-number" in _labels(r)
-
-def test_cl_finds_defstruct():
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    labels = _labels(r)
-    assert "request-stats" in labels
-    # defstruct with options form: (defstruct (name ...) ...)
-    assert "connection" in labels
-
-def test_cl_finds_defvar_defparameter_defconstant():
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    labels = _labels(r)
-    assert "*active-connections*" in labels
-    assert "*default-port*" in labels
-    assert "+max-headers+" in labels
-
-def test_cl_finds_define_condition():
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    assert "server-error" in _labels(r)
-
-def test_cl_finds_custom_definer():
-    """The def-prefix heuristic should catch definline / definline-maybe."""
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    labels = _labels(r)
-    # definline-maybe and definline should produce function-style nodes
-    assert "header=()" in labels
-    assert "header<()" in labels
-
-def test_cl_custom_definer_in_call_graph():
-    """Functions defined via custom definers should appear in the call graph."""
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    calls = _call_pairs(r)
-    # compare-headers calls header= and header< (defined via definline-maybe / definline)
-    assert any("compare-headers" in src and "header=" in tgt for src, tgt in calls)
-    assert any("compare-headers" in src and "header<" in tgt for src, tgt in calls)
-
-def test_cl_operator_names_disambiguated():
-    """upi=, upi<, upi> must produce distinct ids (operator chars matter)."""
-    r = extract_commonlisp(FIXTURES / "sample.lisp")
-    # header= and header< must have different ids
-    eq_ids = [n["id"] for n in r["nodes"] if n["label"] == "header=()"]
-    lt_ids = [n["id"] for n in r["nodes"] if n["label"] == "header<()"]
-    assert len(eq_ids) == 1
-    assert len(lt_ids) == 1
-    assert eq_ids[0] != lt_ids[0]
-
-def test_cl_default_value_not_treated_as_definition():
-    """The def-prefix heuristic must not match denylisted symbols."""
-    import tempfile
-    code = "(in-package :cl-user)\n(default-value foo)\n"
-    with tempfile.NamedTemporaryFile(mode='w', suffix='.lisp', delete=False) as f:
-        f.write(code)
-        path = Path(f.name)
-    try:
-        r = extract_commonlisp(path)
-        labels = _labels(r)
-        # default-value is denylisted, shouldn't create a "foo" node
-        assert "foo" not in labels
-        assert "foo()" not in labels
-    finally:
-        path.unlink()
-
-def test_cl_defs_inside_wrapper_macro():
-    """Definitions nested inside wrapper macros like (optimizing ...) or
-    (eval-when ...) must be extracted. Many CL codebases wrap hot-path
-    inline functions in application-specific macros."""
-    import tempfile
-    code = """(in-package :cl-user)
-(optimizing
- (definline-maybe packet-type (p) (aref p 0))
- (definline-maybe set-packet-type (p v) (setf (aref p 0) v)))
-(eval-when (:compile-toplevel :load-toplevel :execute)
-  (defun helper () 42))
-(progn
-  (defun progn-def () 'ok))
-"""
-    with tempfile.NamedTemporaryFile(mode='w', suffix='.lisp', delete=False) as f:
-        f.write(code)
-        path = Path(f.name)
-    try:
-        r = extract_commonlisp(path)
-        labels = _labels(r)
-        assert "packet-type()" in labels
-        assert "set-packet-type()" in labels
-        assert "helper()" in labels
-        assert "progn-def()" in labels
-    finally:
-        path.unlink()
-
-def test_cl_defs_inside_reader_conditional():
-    """#+feature / #-feature reader conditionals wrap their guarded form
-    in an include_reader_macro AST node, which the walker must descend
-    into to find the nested definition."""
-    import tempfile
-    code = """(in-package :cl-user)
-#+little-endian
-(definline-maybe byte-hash= (a b) (eq a b))
-#-sbcl
-(defun only-on-non-sbcl () 1)
-"""
-    with tempfile.NamedTemporaryFile(mode='w', suffix='.lisp', delete=False) as f:
-        f.write(code)
-        path = Path(f.name)
-    try:
-        r = extract_commonlisp(path)
-        labels = _labels(r)
-        assert "byte-hash=()" in labels
-        assert "only-on-non-sbcl()" in labels
-    finally:
-        path.unlink()
-
-def test_cl_defparameter_string_value_not_docstring():
-    """For defvar/defparameter/defconstant, a string literal in the VALUE
-    position must not be wrongly captured as a docstring node."""
-    import tempfile
-    code = '''(in-package :cl-user)
-(defparameter *config-path* "/etc/app/config")
-(defvar *greeting* "hello world")
-'''
-    with tempfile.NamedTemporaryFile(mode='w', suffix='.lisp', delete=False) as f:
-        f.write(code)
-        path = Path(f.name)
-    try:
-        r = extract_commonlisp(path)
-        labels = _labels(r)
-        assert "*config-path*" in labels
-        assert "*greeting*" in labels
-        # The string VALUES must not show up as rationale nodes
-        assert not any("/etc/app/config" in l for l in labels)
-        assert not any("hello world" in l for l in labels)
-    finally:
-        path.unlink()
-
-
-# ── extract() dispatch ────────────────────────────────────────────────────────
-
-def test_extract_dispatches_all_languages():
-    files = [
-        FIXTURES / "sample.py",
-        FIXTURES / "sample.ts",
-        FIXTURES / "sample.go",
-        FIXTURES / "sample.rs",
-        FIXTURES / "sample.lisp",
-    ]
-    r = extract(files)
-    source_files = {n["source_file"] for n in r["nodes"] if n["source_file"]}
-    assert any("sample.py" in f for f in source_files)
-    assert any("sample.ts" in f for f in source_files)
-    assert any("sample.go" in f for f in source_files)
-    assert any("sample.rs" in f for f in source_files)
-    assert any("sample.lisp" in f for f in source_files)
-
-
-# ── Cache ─────────────────────────────────────────────────────────────────────
-
-def test_cache_hit_returns_same_result(tmp_path):
-    src = FIXTURES / "sample.py"
-    dst = tmp_path / "sample.py"
-    dst.write_bytes(src.read_bytes())
-
-    r1 = extract([dst])
-    r2 = extract([dst])
-    assert len(r1["nodes"]) == len(r2["nodes"])
-    assert len(r1["edges"]) == len(r2["edges"])
-
-def test_cache_miss_after_file_change(tmp_path):
-    dst = tmp_path / "a.py"
-    dst.write_text("def foo(): pass\n")
-    r1 = extract([dst])
-
-    dst.write_text("def foo(): pass\ndef bar(): pass\n")
-    r2 = extract([dst])
-    # bar() should appear in the second result
-    labels2 = [n["label"] for n in r2["nodes"]]
-    assert any("bar" in l for l in labels2)
